ml/requirements.txt
pandas
lightgbm
snowflake-connector-python
scikit-learn
joblib
pyarrow